Refactoring remote package
