+++ tests/test_xyzzyx.py
@@ -1,5 +1,7 @@
 import unittest
-import ndio.remote.neurodata as neurodata
+#import ndio.remote.neurodata as neurodata
+from ndio.remote.data import data
+from ndio.remote.resources import resources
 import ndio.remote.errors
 import numpy
 import random
@@ -11,19 +13,19 @@ class TestXYZZYX(unittest.TestCase):
     def setUp(self):
         self.token_user = '043813bf95d9d5be2bb19448d5a9337db086c559'
         hostname = test_settings.HOSTNAME
-        self.nd = neurodata(self.token_user, hostname=hostname)
-        self.nd_force_chunk = neurodata(self.token_user, hostname=hostname,
+        self.resources = resources(self.token_user, hostname=hostname)
+        self.data_utils = data(self.token_user, hostname=hostname,
                                         chunk_threshold=0)
         dataset_name = 'demo1'
         project_name = 'ndio_demos'
         project_token = 'test_token'
         channel_name = 'image1'
-        self.nd.create_dataset(dataset_name, 100, 100, 100, 1.0, 1.0, 1.0)
-        self.nd.create_project(project_name, dataset_name, 'localhost',
+        self.resources.create_dataset(dataset_name, 100, 100, 100, 1.0, 1.0, 1.0)
+        self.resources.create_project(project_name, dataset_name, 'localhost',
                                1, 1, 'localhost', 'Redis')
-        self.nd.create_project_token(dataset_name, project_name,
+        self.resources.create_project_token(dataset_name, project_name,
                                      project_token, 1)
-        self.nd.create_channel(channel_name, project_name,
+        self.resources.create_channel(channel_name, project_name,
                                dataset_name, 'timeseries',
                                'uint8', 0, 500, 0, 0, 0)
 
@@ -32,9 +34,9 @@ class TestXYZZYX(unittest.TestCase):
         project_name = 'ndio_demos'
         project_token = 'test_token'
         channel_name = 'image1'
-        self.nd.delete_channel(channel_name, project_name, dataset_name)
-        self.nd.delete_project(project_name, dataset_name)
-        self.nd.delete_dataset(dataset_name)
+        self.resources.delete_channel(channel_name, project_name, dataset_name)
+        self.resources.delete_project(project_name, dataset_name)
+        self.resources.delete_dataset(dataset_name)
 
     def test_post_get_no_chunk(self):
         token = 'test_token'
@@ -42,11 +44,12 @@ class TestXYZZYX(unittest.TestCase):
         cutout = numpy.zeros((10, 10, 10)).astype(int)
         zind = random.randint(0, 4)
         cutout[2, 3, zind] = random.randint(100, 200)
-        self.nd.post_cutout(token, channel, 20, 20, 20, cutout, resolution=0)
-        pulldown = self.nd.get_cutout(token, channel,
+        self.data_utils.post_cutout(token, channel, 20, 20, 20, cutout, resolution=0)
+        pulldown = self.data_utils.get_cutout(token, channel,
                                       20, 25, 20, 25, 20, 25, resolution=0)
 
-        self.assertEqual(cutout[2, 3, zind], pulldown[3, zind, 0, 2])
+        #self.assertEqual(cutout[2, 3, zind], pulldown[3, zind, 0, 2])
+        self.assertEqual(cutout[2, 3, zind], pulldown[2,3,zind])
 
     def test_post_get_chunk(self):
         token = 'test_token'
@@ -54,9 +57,9 @@ class TestXYZZYX(unittest.TestCase):
         cutout = numpy.zeros((10, 10, 10)).astype(int)
         zind = random.randint(0, 4)
         cutout[2, 3, zind] = random.randint(100, 200)
-        self.nd_force_chunk.post_cutout(token, channel,
+        self.data_utils.post_cutout(token, channel,
                                         20, 20, 20, cutout, resolution=0)
-        pulldown = self.nd_force_chunk.get_cutout(token, channel,
+        pulldown = self.data_utils.get_cutout(token, channel,
                                                   20, 25, 20, 25, 20, 25,
                                                   resolution=0)
         self.assertEqual(cutout[2, 3, zind], pulldown[2, 3, zind])
